#8 fix(banner): revise windows path canon
