+++ Cargo.toml
@@ -29,3 +29,6 @@ tower-http = { version = "0.6.6", features = ["trace"] }
 tracing = "0.1.41"
 tracing-subscriber = "0.3.20"
 unicode-width = "0.2.2"
+
+[target.'cfg(windows)'.dependencies]
+dunce = "1.0.5"
